Snapshot of sierra.base.git from commit 2e56922cb7316c04679be1d29fabfbbc01669f27
From repository at sierra-git.sandia.gov:/git/sierra.base.git

At commit:
commit 2e56922cb7316c04679be1d29fabfbbc01669f27
Author: Brent Perschbacher <bmpersc@sandia.gov>
Date:   Wed Jun 28 15:39:39 2017 -0600

    SEACAS: comment out slice from Trilinos cmake dependencies file

    The slice directory doesn't exist which make Trilinos fail to configure
    when it tries to find it.

+++ packages/seacas/libraries/ioss/src/Ioss_StructuredBlock.C
@@ -49,7 +49,7 @@ namespace {
 
   int sign(int value) { return value < 0 ? -1 : 1; }
 
-  int del(int v1, int v2) { return (std::abs(v1) == std::abs(v2)); }
+  int del(int v1, int v2) { return static_cast<int>(std::abs(v1) == std::abs(v2)); }
 
 } // namespace
